Uncoupled CommentInput from Contracts
